fix: add game stat before main training
